# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/roles": {
            "get": {
                "security": [
                    {
                        "ServiceName": []
                    },
                    {
                        "ServiceKey": []
                    }
                ],
                "description": "Get a paginated list of roles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "List roles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by name",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.ListOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ServiceName": []
                    },
                    {
                        "ServiceKey": []
                    }
                ],
                "description": "Create a new role with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Create a new role",
                "parameters": [
                    {
                        "description": "Role info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.CreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.CreateOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/roles/{id}": {
            "delete": {
                "security": [
                    {
                        "ServiceName": []
                    },
                    {
                        "ServiceKey": []
                    }
                ],
                "description": "Delete a role by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Delete a role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.DeleteOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.CreateInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "description": "Descricao da role",
                    "type": "string",
                    "maxLength": 500
                },
                "name": {
                    "description": "Nome da role",
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.CreateOutput": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Timestamp no formato RFC3339",
                    "type": "string"
                },
                "id": {
                    "description": "ID gerado (UUID v7)",
                    "type": "string"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.DeleteOutput": {
            "type": "object",
            "properties": {
                "deleted_at": {
                    "description": "Timestamp da delecao",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.ListOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.RoleOutput"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.PaginationOutput"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.PaginationOutput": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_role_dto.RoleOutput": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_infrastructure_web_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "example": "error description"
                        }
                    }
                }
            }
        }
    }
}